Remove the default karaoke skin. Because need to read the config from the karaoke skin. So use tricky way to fix the test case broken. Technically, should be better to use the skin transformer ;_;
